fixed bugs in edit order and delete order. Should now work as intended

+++ Web/backend/views/views.py
@@ -42,7 +42,6 @@ from django.utils.timezone import now
 
 class Article(APIView):
     '''Article view.'''
-    #authentication_classes = (TokenAuthentication,)
     # Dependencies are injected, I hope that we will be able to mock
     # (i.e. make stubs of) these for testing
 
@@ -87,7 +86,6 @@ class Article(APIView):
 
 class Group(APIView):
     '''Group.'''
-    #authentication_classes = (TokenAuthentication,)
     # Dependencies are injected, I hope that we will be able to mock
     # (i.e. make stubs of) these for testing
 
@@ -111,7 +109,6 @@ class Group(APIView):
 
 class Storage(View):
     '''Storage view.'''
-    #authentication_classes = (TokenAuthentication,)
     # Dependencies are injected, I hope that we will be able to mock
     # (i.e. make stubs of) these for testing
 
@@ -151,7 +148,6 @@ class Storage(View):
 
 class NearbyStorages(APIView):
     '''Get nearby storages.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps, *args):
@@ -180,7 +176,6 @@ class NearbyStorages(APIView):
 
 class Compartments(APIView):
     '''Compartment view.'''
-    #authentication_classes = (TokenAuthentication,)
     # Dependencies are injected, I hope that we will be able to mock
     # (i.e. make stubs of) these for testing
 
@@ -257,7 +252,6 @@ class Compartments(APIView):
 
 class Order(APIView):
     '''Order endpoint handling all request from /orders'''
-    #authentication_classes = (TokenAuthentication, )
 
     @si.inject
     def __init__(self, _deps, *args):
@@ -324,7 +318,6 @@ class Order(APIView):
 
 class OrderId(APIView):
     '''Order Endpoint which handles all request coming from /orders/id'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps, *args):
@@ -350,7 +343,10 @@ class OrderId(APIView):
         except:
             return Response({'JSON payload not correctly formatted'}, status=status.HTTP_400_BAD_REQUEST)
 
-        current_state = self.order_service.get_order_by_id(id).order_state
+        order = self.order_service.get_order_by_id(id)
+        if order is None:
+            return Response({'Order does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        current_state = order.order_state
 
         if current_state == 'delivered' and order_state == 'delivered' or current_state == 'order placed' and order_state == 'order placed':
             return Response({'This order has already been handled'}, status=status.HTTP_400_BAD_REQUEST)
@@ -374,6 +370,10 @@ class OrderId(APIView):
 
     def delete(self, request, id):
         '''Delete order func'''
+        orders = self.order_service.get_orders()
+        if orders is None:
+            return Response({'Order does not exist. Cannot be deleted'}, status=status.HTTP_404_NOT_FOUND)
+
         deleted_order = self.order_service.delete_order(id)
         if deleted_order is None:
             return Response({'Order deletion failed'}, status=status.HTTP_400_BAD_REQUEST)
@@ -455,7 +455,6 @@ class LoginWithBarcodeOrNfc(APIView):
 
 class SeeAllStorages(View):
     '''See all storages view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps, *args):
@@ -497,7 +496,6 @@ class SeeAllStorages(View):
 
 class AddInputUnit(View):
     '''Add input unit view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -533,7 +531,6 @@ class AddInputUnit(View):
 
 class GetUserTransactions(View):
     '''Get user transactions view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -564,7 +561,6 @@ class GetUserTransactions(View):
 
 class ReturnUnit(View):
     '''Return unit view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -595,7 +591,6 @@ class ReturnUnit(View):
 
 class Transactions(APIView):
     '''Transactions API view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -691,7 +686,6 @@ class Transactions(APIView):
 
 class TransactionsById(APIView):
     '''Get transaction by ID view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -756,7 +750,6 @@ class GetStorageValue(View):
 
 class GetStorageCost(APIView):
     '''Get storage cost API view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps, *args):
@@ -795,7 +788,6 @@ class GetArticleAlternatives(View):
        their attributes. If an article id and a storage id is entered, the
        method returns the id for alternative articles and the amount of
        the alternative articles in that storage'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -841,7 +833,6 @@ class GetArticleAlternatives(View):
 # FR 8.1 start #
 class SearchForArticleInStorages(View):
     '''Search for article in storages view.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -914,7 +905,6 @@ class SearchForArticleInStorages(View):
 
 class ArticleToCompartmentByQRcode(APIView):
     '''Change Article linked to Compartment by using QR code.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps):
@@ -1069,7 +1059,6 @@ class MoveArticle(APIView):
     
 class GetArticles(APIView):
     '''Get articles according to lioNr, name or storageId.'''
-    #authentication_classes = (TokenAuthentication,)
 
     @si.inject
     def __init__(self, _deps, *args):
